bridge/requirements.txt
fastapi==0.115
google-cloud-firestore==2.20
google-cloud-storage==3.1
